security/requirements.txt
Flask
flask_restful
pymongo
pymongo[srv]
python-decouple
requests
mongomock